Moved ItemDetails to it's own file
